requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
passlib[argon2]
email-validator
python-dotenv
